# Conflicts:
#	src/main/java/com/ysoft/dctrl/editor/importer/ImportRunner.java
#	src/main/java/com/ysoft/dctrl/editor/importer/YieldImportRunner.java

+++ src/main/java/com/ysoft/dctrl/editor/importer/AbstractModelImporter.java
@@ -48,7 +48,7 @@ public abstract class AbstractModelImporter<R> implements ModelImporter<R> {
 
     @Override
     public double getProgress() {
-        return size != 0 ? bytesRead*100/size : 0;
+        return size != 0 ? bytesRead*100/ (double) size : 0;
     }
 
     @Override
